int to long

+++ 0x13-bit_manipulation/5-flip_bits.c
@@ -11,7 +11,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int xoint, counter = 0;
+	unsigned long int xoint, counter = 0;
 
 	xoint = (n ^ m);
